refactor(util): Move VersionExt back into main cargo crate
This is a partial revert of 2b2502f7fc4bb8b177f6378647bd65e8e4a00200

These were pulled out with the idea of being a place to house
`PartialVersion` for `util_schemas` to use.
Since then, we went with a `util_schemas::core` in #13128,
meaning we can hold off on creating yet another crate for us to manage.
